+++ app/src/main/AndroidManifest.xml
@@ -34,9 +34,6 @@
         <activity
             android:name=".views.UserPanel"
             android:parentActivityName=".views.NoticeList" />
-        <activity
-            android:name=".views.NoticeAdd"
-            android:parentActivityName=".views.NoticeList" />
     </application>
 
 </manifest>
