+++ app/src/main/java/com/example/teamyuml/decentworkmobile/fragments/NoticeForm.java
@@ -1,8 +1,11 @@
-package com.example.teamyuml.decentworkmobile.views;
+package com.example.teamyuml.decentworkmobile.fragments;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import android.support.annotation.NonNull;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
 import android.view.View;
+import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.EditText;
 import android.widget.Spinner;
@@ -25,37 +28,36 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+public class NoticeForm extends Fragment implements View.OnClickListener {
 
-/**
- * Class responsible for adding notice
- * Parsing data to JSON and sending to the server
- */
-public class NoticeAdd extends AppCompatActivity implements View.OnClickListener {
+    private Spinner noticeCity;
+    private View v;
+    private int dropDownLayout = android.R.layout.simple_dropdown_item_1line;
     private EditText noticeTitle;
     private EditText noticeDescription;
-    private Spinner noticeCity;
     private final String NOTICE_ADD_URL = VolleyInstance.getBaseUrl() + "/engagments/engagments/";
-    private int dropDownLayout = android.R.layout.simple_dropdown_item_1line;
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_engagment_form);
-        noticeTitle = findViewById(R.id.title);
-        noticeDescription = findViewById(R.id.description);
-        findViewById(R.id.add_notice_btn).setOnClickListener(this);
-        setCitySpinner();
     }
 
-    /**
-     * Makes request to server and add notice if success or shows
-     * error when failure.
-     */
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        v = inflater.inflate(R.layout.notice_form, container, false);
+        noticeTitle = v.findViewById(R.id.title);
+        noticeDescription = v.findViewById(R.id.description);
+        v.findViewById(R.id.add_notice_btn).setOnClickListener(this);
+        setCitySpinner();
+        return v;
+    }
+
     @Override
     public void onClick(View v) {
         try {
             JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(
-                Request.Method.POST, NOTICE_ADD_URL, addParams(), new Response.Listener<JSONObject>() {
+                    Request.Method.POST, NOTICE_ADD_URL, addParams(), new Response.Listener<JSONObject>() {
                 @Override
                 public void onResponse(JSONObject response) {
                     //TODO ADD INTENT TO USER NOTICES
@@ -63,17 +65,17 @@ public class NoticeAdd extends AppCompatActivity implements View.OnClickListener
             }, new Response.ErrorListener() {
                 @Override
                 public void onErrorResponse(VolleyError error) {
-                    ErrorHandler.errorHandler(error, NoticeAdd.this);
+                    ErrorHandler.errorHandler(error, getActivity());
                 }
             }) {
                 public Map<String, String> getHeaders() {
-                    return UserAuth.authorizationHeader(NoticeAdd.this);
+                    return UserAuth.authorizationHeader(getActivity());
                 }
             };
 
-            VolleyInstance.getInstance(this).addToRequestQueue(jsonObjectRequest, "Create notice");
+            VolleyInstance.getInstance(getActivity()).addToRequestQueue(jsonObjectRequest, "Create notice");
         } catch (JSONException e) {
-            Toast.makeText(this, "Nie można sparsować parametrów", Toast.LENGTH_LONG).show();
+            Toast.makeText(getActivity(), "Nie można sparsować parametrów", Toast.LENGTH_LONG).show();
         }
     }
 
@@ -98,9 +100,9 @@ public class NoticeAdd extends AppCompatActivity implements View.OnClickListener
      * Initializes spinner with cities.
      */
     private void setCitySpinner() {
-        noticeCity = findViewById(R.id.city);
+        noticeCity = v.findViewById(R.id.city);
         ArrayAdapter<String> arrayAdapter = new ArrayAdapter<>(
-            NoticeAdd.this, dropDownLayout, populateCities());
+                getActivity(), dropDownLayout, populateCities());
         arrayAdapter.setDropDownViewResource(dropDownLayout);
         noticeCity.setAdapter(arrayAdapter);
     }
@@ -118,4 +120,3 @@ public class NoticeAdd extends AppCompatActivity implements View.OnClickListener
         return cities;
     }
 }
-
